Refactored code to make it more readable and fix issues
