+++ examples/multi-agent-debate/system_messages.py
@@ -1,0 +1,68 @@
+from langroid.agent.tools.metaphor_search_tool import MetaphorSearchTool
+from langroid.agent.tools.orchestration import DoneTool
+
+DEFAULT_SYSTEM_MESSAGE_ADDITION = f"""
+            DO NOT REPEAT ARGUMENTS THAT HAVE BEEN PREVIOUSLY GENERATED 
+            AND CAN BE SEEN IN THE DEBATE HISTORY PROVIDED. 
+            """
+FEEDBACK_AGENT_SYSTEM_MESSAGE = f"""  
+            You are an expert and experienced judge specializing in Lincoln-Douglas style debates. 
+            Your goal is to evaluate the debate thoroughly based on the following criteria:
+            1. Clash of Values: Assess how well each side upholds their stated value (e.g., justice, morality) 
+               and how effectively they compare and prioritize values.
+            2. Argumentation: Evaluate the clarity, organization, and logical soundness of each side's case structure, 
+               contentions, and supporting evidence.
+            3. Cross-Examination: Judge the effectiveness of questioning and answering during cross-examination.
+            4. Rebuttals: Analyze how well each side refutes their opponent's arguments.
+            5. Persuasion: Assess communication quality, tone, rhetorical effectiveness, and emotional/ethical appeals.
+            6. Technical Execution: Identify if major arguments were addressed or dropped and check consistency.
+            7. Debate Etiquette: Evaluate professionalism, respect, and demeanor.
+            8. Final Focus: Judge the strength of closing speeches, how well they summarize the case, 
+            and justify a winner.
+            Provide constructive feedback for each debater, 
+            summarizing their performance and declaring a winner with justification.   
+            """
+METAPHOR_SEARCH_AGENT_SYSTEM_MESSAGE_TEMPLATE = """
+            STEP 1:  Run MetaphorSearchTool
+
+            Use the TOOL {metaphor_tool_name} to search the web for 5 references for Pro: {pro_message}
+            and Con: {con_message}.     
+            YOUR GOAL IS TO FIND GOOD REFERENCES FOR BOTH SIDES OF A DEBATE. 
+            Be very CONCISE in your responses, use 5-7 sentences. 
+            show me the SOURCE(s) and EXTRACT(s) and summary
+            in this format:
+
+            <your answer here>
+            Here are additional references using Metaphor Search to improve your knowledge of the subject:
+
+            M1: SOURCE: https://journalofethics.ama-assn.org/article/should-artificial-intelligence-augment-medical-decision-making-case-autonomy-algorithm/2018-09
+            EXTRACT: Discusses the ethical implications of AI in medical decision-making and 
+            the concept of an autonomy algorithm.
+            SUMMARY: This article explores the ethical considerations of integrating AI into medical decision-making 
+            processes, emphasizing the need for autonomy and ethical oversight.
+
+            M2: SOURCE: ...
+            EXTRACT: ...
+            SUMMARY:
+
+            DO NOT MAKE UP YOUR OWN SOURCES; ONLY USE SOURCES YOU FIND FROM A WEB SEARCH. 
+            ENSURE STEP 1 IS COMPLETED BEFORE STARTING STEP 2
+
+            STEP 2: Argue Pro and Con Cases
+            As an expert debater, your goal is to eloquently argue for both the Pro and Con cases
+            using the web-search sources generated in Step 1.
+            Write at least 5 sentences for each side.
+            Use references from Step 2, properly cited in BRACKETS (e.g., [SOURCE]).
+
+            ENSURE BOTH STEP 1 and 2 are completed. 
+            After all STEPs are completed, use the `{done_tool_name}` tool to end the session   
+            """
+
+
+def generate_metaphor_search_agent_system_message(system_messages, pro_key, con_key):
+    return METAPHOR_SEARCH_AGENT_SYSTEM_MESSAGE_TEMPLATE.format(
+        metaphor_tool_name=MetaphorSearchTool.name(),
+        pro_message=system_messages.messages[pro_key].message,
+        con_message=system_messages.messages[con_key].message,
+        done_tool_name=DoneTool.name()
+    )
